feat: change repository domain id type Integer to Long

+++ src/main/java/com/example/spring_boot_jwt_boilerplate/repository/KanbanBoardRepository.java
@@ -3,5 +3,5 @@ package com.example.spring_boot_jwt_boilerplate.repository;
 import com.example.spring_boot_jwt_boilerplate.domain.kanban.KanbanBoard;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface KanbanBoardRepository extends JpaRepository<KanbanBoard, Integer> {
+public interface KanbanBoardRepository extends JpaRepository<KanbanBoard, Long> {
 }
