+++ src/main/java/com/example/spring_boot_jwt_boilerplate/repository/KanbanTaskRepository.java
@@ -3,5 +3,5 @@ package com.example.spring_boot_jwt_boilerplate.repository;
 import com.example.spring_boot_jwt_boilerplate.domain.task.KanbanTask;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface KanbanTaskRepository extends JpaRepository<KanbanTask, Integer> {
+public interface KanbanTaskRepository extends JpaRepository<KanbanTask, Long> {
 }
